+++ serverless-functions/ReadSourceData/function.json
@@ -1,0 +1,19 @@
+{
+  "scriptFile": "__init__.py",
+  "bindings": [
+    {
+      "name": "blob",
+      "type": "blobTrigger",
+      "direction": "in",
+      "path": "source-data/{name}",
+      "connection": "PhiStorage"
+    },
+    {
+      "type": "serviceBus",
+      "direction": "out",
+      "connection": "AzureServiceBusConnectionString",
+      "name": "queue",
+      "queueName": "%ServiceBusQueueName%"
+    }
+  ]
+}
