--- serverless-functions/read_source_data/requirements.txt
@@ -1,15 +1,0 @@
-cloudevents==1.6.1
-Flask==2.2.2
-functions-framework==3.2.0
-google-api-core==2.8.2
-google-auth==2.11.0
-google-cloud==0.34.0
-google-cloud-core==2.3.2
-google-cloud-pubsub==2.13.6
-google-cloud-storage==2.5.0
-google-crc32c==1.3.0
-google-resumable-media==2.3.3
-googleapis-common-protos==1.56.4
-grpc-google-iam-v1==0.12.4
-phdi==0.1.0.dev1
-phdi_cloud_function_utils @ git+https://github.com/CDCgov/phdi-google-cloud@main#subdirectory=cloud-functions/phdi_cloud_function_utils
