+++ serverless-functions/requirements.txt
@@ -1,0 +1,9 @@
+azure-functions
+azure-identity
+azure-storage-blob
+hl7
+phdi
+requests
+smartystreets_python_sdk
+urllib3
+typer
