+++ serverless-functions/requirements_dev.txt
@@ -1,0 +1,1 @@
+pytest
